Move some classes to separate file
